refactor: move statics from Projectile
