+++ src/main/java/org/runejs/OldEngine/MapDecompressor.java
@@ -5,7 +5,6 @@ package org.runejs.OldEngine;/*
  */
 
 import org.runejs.client.Class26;
-import org.runejs.client.Projectile;
 import org.runejs.client.Landscape;
 import org.runejs.client.MovedStatics;
 import org.runejs.client.media.renderable.actor.Npc;
@@ -134,7 +133,7 @@ public class MapDecompressor {
             class20 = groundData[plane];
         }
         if(localY > 0 && localX > 0 && localY < 103 && localX < 103) {
-            Projectile.addObject(objectId, localX, localY, z, rotation, type, Npc.currentScene, class20);
+            MovedStatics.addObject(objectId, localX, localY, z, rotation, type, Npc.currentScene, class20);
         }
     }
